Moving some methods from DynamoModel to the DummyNode class.
