+++ src/DynamoCore/Graph/Workspaces/WorkspaceModel.cs
@@ -1866,71 +1866,6 @@ namespace Dynamo.Graph.Workspaces
         }
 
         /// <summary>
-        /// This will return the NodeModel for a dummy node after the package(that is loaded) resolves the node. 
-        /// </summary>
-        /// <param name="json"></param>
-        /// <param name="libraryServices"></param>
-        /// <param name="factory"></param>
-        /// <param name="isTestMode"></param>
-        /// <param name="manager"></param>
-        public static NodeModel GetNodeModelForDummyNode(string json, LibraryServices libraryServices,
-                 NodeFactory factory, bool isTestMode, CustomNodeManager manager)
-        {
-            var settings = new JsonSerializerSettings
-            {
-                Error = (sender, args) =>
-                {
-                    args.ErrorContext.Handled = true;
-                    Console.WriteLine(args.ErrorContext.Error);
-                },
-                ReferenceLoopHandling = ReferenceLoopHandling.Ignore,
-                TypeNameHandling = TypeNameHandling.Auto,
-                Formatting = Newtonsoft.Json.Formatting.Indented,
-                Culture = CultureInfo.InvariantCulture,
-                Converters = new List<JsonConverter>{
-                        new NodeReadConverter(manager, libraryServices, factory, isTestMode),
-                        new TypedParameterConverter()
-                    },
-                ReferenceResolverProvider = () => { return new IdReferenceResolver(); }
-            };
-
-            var result = SerializationExtensions.ReplaceTypeDeclarations(json, true);
-            var nodeModel = JsonConvert.DeserializeObject<NodeModel>(result, settings);
-            
-            return nodeModel;
-        }
-
-        /// <summary>
-        /// This will set the dummy node's node view properties to the resolved node
-        /// </summary>
-        /// <param name="dummyNode"></param>
-        /// <param name="resolvedNode"></param>
-        public void SetNodeViewDataOnResolvedNode(NodeModel dummyNode, NodeModel resolvedNode)
-        {
-            if (dummyNode == null || resolvedNode == null)
-                return;
-
-            if (resolvedNode != null)
-            {
-                resolvedNode.X = dummyNode.X;
-                resolvedNode.Y = dummyNode.Y;
-                resolvedNode.IsFrozen = dummyNode.IsFrozen;
-                resolvedNode.IsSetAsInput = dummyNode.IsSetAsInput;
-                resolvedNode.IsSetAsOutput = dummyNode.IsSetAsOutput;
-
-                // NOTE: The name needs to be set using UpdateValue to cause the view to update
-                resolvedNode.UpdateValue(new UpdateValueParams("Name", dummyNode.Name));
-                resolvedNode.UpdateValue(new UpdateValueParams("IsVisible", dummyNode.IsVisible.ToString()));
-            }
-            else
-            {
-                this.Log(string.Format("This graph has a node with id:{0} and name:{1}, but it could not be resolved",
-                                resolvedNode.GUID, resolvedNode.Name)
-                                , WarningLevel.Moderate);
-            }
-        }
-
-        /// <summary>
         /// Updates a workspace model with extra view information. When loading a workspace from JSON,
         /// the data is split into two parts, model and view. This method sets the view information.
         /// </summary>
